Updated to use a d3.js example
jQuery has been removed from now since it's not really needed for the
example.

d3.js has been downloaded to make the example work offline.

Data is generated randomly in the javascript side, but the idea is to
generate the data in the python side and send it to the browser widget.