Added clarification regarding loop
I tried to use a priority queue to handle gtk/webkit messages and also
to enqueue a 'quit' message, so that the signal handler uses the same
queue. However, when the main thread is waiting for the queue, the
signal handler is not called when pressing Ctrl-C until a message in the
queue is received.